config loader cleaning
