+++ scraper/src/tests/config_loader/abstract.py
@@ -12,13 +12,15 @@ def config(additional_config={}):
         'custom_settings': 'custom_settings',
         'hash_strategy': 'hash_strategy',
         'index_name': 'index_name',
-        'index_prefix': 'index_prefix',
         'selectors': [],
         'selectors_exclude': [],
         'start_urls': ['http://www.starturl.com/'],
         'stop_urls': ['http://www.stopurl.com/'],
         'strategy': 'strategy',
-        'strip_chars': 'strip_chars'
+        'strip_chars': 'strip_chars',
+        'js_render': False,
+        'js_wait': 0,
+        'use_anchors': False
     }
     final_config = base_config.copy()
     final_config.update(additional_config)
