+++ scraper/src/tests/config_loader/basic_test.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 import os
 
-from ...config_loader import ConfigLoader
+from ...config.config_loader import ConfigLoader
 from .abstract import config
 import pytest
