+++ scraper/src/tests/config_loader/domains_test.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-from ...config_loader import ConfigLoader
+from ...config.config_loader import ConfigLoader
 from .abstract import config
 
 class TestDomains:
@@ -7,7 +7,7 @@ class TestDomains:
         """ Should populate allowed_domains from start_urls """
         # Given
         c = config({
-            'start_urls': 'http://www.foo.bar/',
+            'start_urls': ['http://www.foo.bar/'],
             'stop_urls': [],
             'allowed_domains': None
         })
@@ -22,8 +22,8 @@ class TestDomains:
         """ Should populate allowed_domains from both start and stop urls """
         # Given
         c = config({
-            'start_urls': 'http://www.foo.bar/',
-            'stop_urls': 'http://www.algolia.com/',
+            'start_urls': ['http://www.foo.bar/'],
+            'stop_urls': ['http://www.algolia.com/'],
             'allowed_domains': None
         })
 
@@ -37,7 +37,7 @@ class TestDomains:
         """ Should populate a list of unique domains """
         # Given
         c = config({
-            'start_urls': 'http://www.foo.bar/',
+            'start_urls': ['http://www.foo.bar/'],
             'stop_urls': [
                 'http://www.algolia.com/',
                 'http://www.foo.bar/'
@@ -51,12 +51,12 @@ class TestDomains:
         # Then
         assert actual.allowed_domains == ['www.foo.bar', 'www.algolia.com']
 
-    def test_allowed_domains_accept_single_value(self):
+    def test_not_allowed_domains_accept_single_value(self):
         """ Should allow passing allowed_domains as a string instead of an array
         """
         # Given
         c = config({
-            'allowed_domains': 'www.foo.bar'
+            'allowed_domains': ['www.foo.bar']
         })
 
         # When
