+++ scraper/src/tests/config_loader/open_selenium_browser_test.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-from ...config_loader import ConfigLoader
+from ...config.config_loader import ConfigLoader
 from .abstract import config
