+++ scraper/src/tests/config_loader/stop_urls_test.py
@@ -1,8 +1,7 @@
 # coding: utf-8
-import json
-import os
+import pytest
 
-from ...config_loader import ConfigLoader
+from ...config.config_loader import ConfigLoader
 from .abstract import config
 
 
@@ -11,7 +10,7 @@ class TestStopUrls:
         """ Allow passing stop_urls as string instead of array """
         # Given
         c = config({
-            'stop_urls': 'www.foo.bar'
+            'stop_urls': ['www.foo.bar']
         })
 
         # When
@@ -23,19 +22,12 @@ class TestStopUrls:
     def test_stop_urls_is_not_mandatory(self):
         """ Allow not passing stop_urls """
         # Given
-        conf = {
-            'allowed_domains': 'allowed_domains',
-            'api_key': 'api_key',
-            'app_id': 'app_id',
-            'index_name': 'index_name',
-            'index_prefix': 'index_prefix',
-            'selectors': [],
-            'selectors_exclude': [],
+        conf = config({
             'start_urls': ['http://www.starturl.com/']
-        }
+        })
 
-        # When
-        actual = ConfigLoader(json.dumps(conf))
-
-        # Then
-        assert actual.stop_urls == []
+        with pytest.raises(Exception) as excinfo:
+            # When
+            ConfigLoader(conf)
+            # Then
+            assert 'start_urls should be list' in str(excinfo.value)
